pyproject.toml
[project]
name = "cfg-neuralese"
version = "0.1.0"
description = "CFG-Evolving Neuralese Communication Protocol"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "lark>=1.1.0",
    "matplotlib>=3.5.0",
    "numpy>=1.20.0",
    "pytest>=8.4.1",
]
